refactor button addForum
