+++ app/src/main/java/com/integrapp/integrapp/DocuFragment.java
@@ -27,7 +27,6 @@ public class DocuFragment extends Fragment {
     private ArrayList<ForumItem> threads = new ArrayList<>();
     private ListView llista;
     private ForumsAdapter forumsAdapter;
-    private FloatingActionButton fab;
 
     public void onCreate (Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -37,20 +36,9 @@ public class DocuFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_forum_docu, container, false);
-        fab = view.findViewById(R.id.fab);
         this.server = Server.getInstance();
         llista = view.findViewById(R.id.llista);
         setInfoForum();
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                FragmentManager fragmentManager = getActivity().getSupportFragmentManager();
-                FragmentTransaction ft = fragmentManager.beginTransaction();
-                ft.replace(R.id.screen_area, new NewForumFragment());
-                ft.addToBackStack(null);
-                ft.commit();
-            }
-        });
         return view;
     }
