+++ app/src/main/java/com/integrapp/integrapp/EntreFragment.java
@@ -25,7 +25,6 @@ public class EntreFragment extends android.support.v4.app.Fragment {
     private ArrayList<ForumItem> threads = new ArrayList<>();
     private ListView llista;
     private ForumsAdapter forumsAdapter;
-    private FloatingActionButton fab;
 
     public void onCreate (Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -36,19 +35,8 @@ public class EntreFragment extends android.support.v4.app.Fragment {
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_forum_entre, container, false);
         this.server = Server.getInstance();
-        fab = view.findViewById(R.id.fab);
         llista = view.findViewById(R.id.llista);
         setInfoForum();
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                FragmentManager fragmentManager = getActivity().getSupportFragmentManager();
-                FragmentTransaction ft = fragmentManager.beginTransaction();
-                ft.replace(R.id.screen_area, new NewForumFragment());
-                ft.addToBackStack(null);
-                ft.commit();
-            }
-        });
         return view;
     }
